When renaming a table column try to fix all indices referencing the column
Suppose you have this database:
CREATE TABLE a(
    x INTEGER,
    y INTEGER
);
CREATE INDEX i ON a(x ASC);

Now, when trying to rename column x of table a to some other name using
the Edit Table dialog will work but index i will be lost during the
process. This is because it needs to be recreated after having modified
the table, however its create statement isn't valid anymore because the
column name x doesn't exist anymore. This is changed by this commit to
significantly increase the chance that renaming a column won't throw an
error, thus making the Edit Table dialog a lot safer to use.