Don't use base table name where the actual table name is appropriate
In the SQL export don't use the base table name of a table when you
should actually use the table name itself. The way SQLite reports its
data these just happen to be the same for tables.

This doesn't change anything when it comes to functionality but is
semantically more correct.