dbhub: Use commit id instead of version number
This changes all the code to use the commit id of a database version
instead of a version number. This means changing the UI, the code for
the communication with dhhub.io, and the internal database for storing
information on downloaded databases.

The last step makes it necessary to delete that database file on your
system, so it will be recreated! Please do that :)