Fix index updating when removing an indexed column from a table
When changing an indexed column from a table using the Edit Table dialog
we try to update the corresponding indices as well. So if a column is
renamed all references to it in the indices are renamed accordingly.
This, however, didn't take into account that you can also remove
columns. While trying to update the column to nothing, the application
would crash. This is fixed by this commit. It doesn't crash anymore and
now removes the column from the index as expected.

See issue #1115.