Fixed inconsistencies in new toolbar buttons
Restored menu entries for 'Open Project' and 'Save Project'. They were
accidentally removed in 8fbf44bffec5c73ad623e6d90ea4667dc0ef94e4

Set the status-tip to all the toolbar buttons for consistency. This
displays a message in the status bar when hovering over the toolbar button,
and (depending on platform), when the menu entry is the current. Set to the
the same text than the tool-tip following previous example.

Both tips for 'Close Database' were added.